refactor(simulation): split NoProgressDetector into focused modules
- Extract types (BattleUnit, BattleState, getUnitName) to types.ts (34 lines)
- Extract snapshot logic (creation, comparison, serialization) to snapshots.ts (137 lines)
- Extract detection algorithm to detection.ts with NoProgressDetectionEngine (303 lines)
- Create barrel export in index.ts with NoProgressDetector wrapper class (39 lines)
- Reduce original file to facade re-export (1 line)

All modules under 400 lines. Follows StateCycleDetector pattern established in TODO 17.
Preserves 100% of public API via facade - zero breaking changes for consumers.
